+++ app/Models/Conversation.php
@@ -33,7 +33,6 @@ use Illuminate\Support\Facades\Auth;
  * @method static \Illuminate\Database\Eloquent\Builder|Conversation whereSenderId($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Conversation whereSenderRead($value)
  * @method static \Illuminate\Database\Eloquent\Builder|Conversation whereUpdatedAt($value)
- * @property-read \Illuminate\Database\Eloquent\Collection<int, \App\Models\Message> $messages
  * @mixin \Eloquent
  */
 class Conversation extends Model