+++ public/assets/js/custom.js
@@ -90,7 +90,7 @@ $('#switch_rides').click(function () {
 
 //MODAL
 //values in array corresponds to modalClass in custom-modal-content.blade.php
-let modals = ['appstoreinfo'];
+let modals = [/*'appstoreinfo',*/ 'driverprofile'];
 $.each(modals, function (index, modalClass) {
     var localStorageItem = 'modalClosed-' + modalClass;
     if (!localStorage.getItem(localStorageItem)) {