+++ resources/views/components/phone-number.blade.php
@@ -1,0 +1,6 @@
+@if($phoneNumber && $isPhoneNumberPublic)
+    <a class="underline"
+       href="tel:{{$phoneNumber}}">{{$phoneNumber}}</a>
+@else
+    {{__('Phone not added')}}
+@endif