+++ resources/views/profile/show.blade.php
@@ -8,11 +8,12 @@
     <div>
         <div class="max-w-7xl mx-auto py-10 sm:px-6 lg:px-8">
             @livewire('profile.update-driver-profile-information-form')
-
             <x-section-border/>
 
             @if (Laravel\Fortify\Features::canUpdateProfileInformation())
-                @livewire('profile.update-profile-information-form')
+                <div class="mt-10 sm:mt-0">
+                    @livewire('profile.update-profile-information-form')
+                </div>
 
                 <x-section-border/>
             @endif