+++ resources/views/ride-requests/my-requests/components/list_result.blade.php
@@ -1,7 +1,4 @@
 <?php
-
-use App\Source\RideRequest\Enum\RideRequestEnum;
-
 /** @var \Illuminate\Contracts\Pagination\LengthAwarePaginator $requests */
 
 /** @var \App\Models\RideRequest $request */
@@ -9,9 +6,6 @@ use App\Source\RideRequest\Enum\RideRequestEnum;
 ?>
 <div
     class="p-6 lg:p-8 bg-white dark:bg-gray-800 dark:bg-gradient-to-bl dark:from-gray-700/50 dark:via-transparent border-b border-gray-200 dark:border-gray-700">
-
-    <h1 class="pb-6 text-4xl dark:text-white">{{$ride->fromPlace->getName()}} - {{$ride->toPlace->getName()}}</h1>
-    <h2 class="pb-6 text-xl dark:text-white">{{__('Departure time')}} {{$ride->getTimeFormatted()}}</h2>
 
     <ul class="divide-y divide-gray-200 dark:divide-gray-700">
         @foreach($requests as $request)
@@ -23,8 +17,7 @@ use App\Source\RideRequest\Enum\RideRequestEnum;
                 <div class="flex items-center space-x-4">
                     <div class="flex-shrink-0">
                         <a class="underline" href="{{user_profile_url($request->passenger->getId())}}">
-                            <img class="w-8 h-8 rounded-full" src="{{$request->passenger->getProfilePhotoUrl()}}"
-                                 alt="Neil image">
+                            <img class="w-8 h-8 rounded-full" src="{{$request->passenger->getProfilePhotoUrl()}}">
                         </a>
                     </div>