+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.17.0] - 2023-04-16
+
+### Added
+
+- Added driver profile
+- Add phone number field
+
 ## [1.16.2] - 2023-04-17
 
 ### Fixed